rename batch -> command
