+++ command/account.go
@@ -1,4 +1,4 @@
-package batch
+package command
 
 import (
 	"flag"
