+++ command/account_test.go
@@ -1,4 +1,4 @@
-package batch
+package command
 
 import (
 	"fmt"
