+++ main.go
@@ -4,7 +4,7 @@ import (
 	"flag"
 	"fmt"
 	"github.com/labstack/gommon/log"
-	"tools/batch"
+	"tools/command"
 	"tools/conf"
 	"tools/db"
 )
@@ -23,9 +23,9 @@ func main() {
 	}
 
 	subCommands := map[string]func(){
-		"AddAccount":   batch.Add,
-		"ShowAccount":  batch.Show,
-		"CopyPassword": batch.CopyPassword,
+		"AddAccount":   command.Add,
+		"ShowAccount":  command.Show,
+		"CopyPassword": command.CopyPassword,
 	}
 
 	command := args[0]
